pyproject.toml
[tool.poetry]
name = "AgriData Relay"
version = "0.1.0"
description = ""
authors = ["Leon Bowie <leon@bowie-co.nz>"]

[tool.poetry.dependencies]
python = "^3.10"
substrate-interface = "^1.1.7"
PyNaCl = "^1.5.0"
asyncio-mqtt = "^0.12.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
